pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "shunya-py"
version = "0.2.0"
description = "Multi-ticker equity panels, JAX alphas, and backtrader backtests with yfinance or Alpaca data."
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "Kaushik Dey" }]
keywords = [
    "alpaca",
    "backtesting",
    "backtrader",
    "equities",
    "finance",
    "jax",
    "quantitative",
    "yfinance",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
    "Typing :: Typed",
]
dependencies = [
    "alpaca-py>=0.43.2",
    "backtrader>=1.9.78.123",
    "finta>=1.3",
    "jax>=0.9.2",
    "jaxlib>=0.9.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pytickersymbols>=1.17.10",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "tqdm>=4.67.3",
    "yfinance>=1.2.0",
]

[project.optional-dependencies]
notebook = ["curl-cffi>=0.6.0", "ipykernel>=7.2.0"]
dev = ["pytest>=8.0", "twine>=6.0"]

[project.urls]
Homepage = "https://github.com/Kaushikdey647/shunya"
Repository = "https://github.com/Kaushikdey647/shunya"
Documentation = "https://github.com/Kaushikdey647/shunya#readme"
Changelog = "https://github.com/Kaushikdey647/shunya/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["shunya"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
